use $dsl->config instead of plugin_settings for plugin2
Any plugins which use this plugin will not work when plugin1's
plugin_config keyword is used and so if we smell plugin2 we get config
via $dsl->config instead. This is needed by
Dancer2::Plugin::Auth::Extensible::Provider::DBIC (and probably some
other plugins - still some work to do).

Thanks to yanick for the pointer.